refactor: Remove target reached from standard aborters (#2912)
Currently the aborters set or get the `targetReached` property which is unwanted from my point of view because the aborters should only decide on a minimal amount of information which is captured in the name. e.g. `PathLimitReached` should not care about if we reached the target but only about the path length and limit