refactor: Assert to be `on surface` for surface functions with free param input (#2932)
This popped up in https://github.com/acts-project/acts/pull/2811 - as this can be a pitfall we are asserting if we are really on surface. This is done via an `assert` which is not evaluated in standard release build

This also fixes a bug in the covariance engine. We used to first do the cov transport and then check if we are actually on surface. I turned these checks around.