fix: TrackStateType printout in wrong order (#2950)
`operator<<(std::ostream& os, ConstTrackStateType t)` prints out the flag bits as a binary number, with bit 0 (`MeasurementFlag`) at the end. This is belied by the label `MPOHMS`, which should be `SMHOPM`.

This PR fixes the label. Another option would be to reverse the order of bit printout. Please comment if you think that would be better.