feat: encapsulate assignment of material interactions to material surfaces (MM1)  (#3015)
This is the first PR in a series that divides the Material Mapping into logical, unit testable modules:

1) Finding intersections with surfaces and associations to volumes (MM2)
2) Assigning material interactions to those intersections (MM1, this PR)
3) Mapping those onto dedicated Surface / Volume Material Mappers (MM3)
4) Steer that by a chained algorithm (MM4)

This PR:

The logic how to walk through material interactions and assign those is currently implemented in both the `SurfaceMaterialMapper` and `VolumeMaterialMapper` and consists of many If/else cases that are rather difficult to follow.
This PR (after consultation with @Corentin-Allaire ) introduces a single logic for Surface based material that allows to define global and local vetos (global, e.g. for volume restrictions) and re-assignment policies for pre/post mapping directives.

All cases are showcased and tested in a set of UnitTests.