refactor: Refactor how the bound state covariance is handled without transport (#3104)
Refactors the internal wirings of `boundState` if no covariance transport is requested. Currently the function might still return a covariance depending on the state of the stepper. This is a workaround to access the covariance of the stepper in case it was transported before.

I refactored this and left the responsibility to the caller to get a covariance in case it was transported before hand.

While doing this I also removed some unnecessary initializations and removed `std::move`s from eigen objects because they do not have an effect.