perf: Try fast pow for `EigenStepper` step size scaling (#3153)
Currently the step scaling calculation has a big impact on the `EigenStepper` performance. In this PR the `pow(x, 0.25)` is approximated with `fastPow` (similar to fast inverse square root) which relies on the bit representation of the floating point number.

The assumption is that we do not really care about the precision of this value but rather that it gives a good approximation for the step size scaling.

Edit: After measuring the performance it seems like there is no measurable improvement. I made this approximation compile time optional for now. See this https://github.com/acts-project/acts/pull/3153#issuecomment-2082073678 for more details

<img width="577" alt="image" src="https://github.com/acts-project/acts/assets/487211/d613caeb-991f-4a89-98fc-bc051be13b7b">

References
- https://martin.ankerl.com/2012/01/25/optimized-approximative-pow-in-c-and-cpp